+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # CHANGELOG
 
-## v0.0.3 - WIP
+A Work in Progress (WIP).
+No tests.
+Not yet ready for use.
+
+## v0.0.3 - 2023-08-04 WIP
 - Make `pan_length` first parameter of `pan_range`.
 - China UnionPay, China T-Union, Diners Club, Diners Club International,
 - JCB, Maestro, Maestro UK, Visa Electron
+- BORICA, Humo, InstaPayment, InterPayment, Mir, Napas, UATP, UzCard
 
-## v0.0.2 - 2023-08-03
+## v0.0.2 - 2023-08-03 WIP
 - Initial release.
 - American Express, Discover, Mastercard, and Visa.